% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_centering_by_residuals.R
\name{addlevelMods}
\alias{addlevelMods}
\title{addlevelMods}
\usage{
addlevelMods(dat)
}
\arguments{
\item{dat}{}
}
\description{
This function adds the high and low moderators to the data as columns in the dataframe.
}
