% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_model.R
\name{basic_co}
\alias{basic_co}
\title{Title
this function basically has no involvement from the partner.}
\usage{
basic_co(data, tau, dim)
}
\arguments{
\item{dim}{}
}
\description{
Title
this function basically has no involvement from the partner.
}
