% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tool.R
\name{chec_data_trellis}
\alias{chec_data_trellis}
\title{chec_data_trellis}
\usage{
chec_data_trellis(dat)
}
\arguments{
\item{dat}{}
}
\description{
This function is used to help visualize whether the data coming into the build_data function is poorly edited. The package doesn't have extensive checking steps in order to keep it is general as possible to the varieties of Dyadic Timeseries data that may exist.
}
\details{
The idea is to use check the plots and return to the raw csv to edit out what is necessary.
}
