% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_actor_partner_cross.R
\name{crossPartners}
\alias{crossPartners}
\title{crossPartners}
\usage{
crossPartners(data)
}
\arguments{
\item{data}{}
}
\value{
a new dataframe taht has a wide format with a person and partenrs data on the same row
}
\description{
This function does the much anticipated actual flipping of the data. Effectively turns a long datastructure in a wide where a specific person's id
has their partner's data also within that row. Very helpful format for
performing fitting and other types of analysis.
}
\details{
In the example below the argument is equalized and estimated data, because
that illustrates that this package uses this function following the 
derivative estimation step.
}
\examples{
crossPartners(equalized.estimated.data)
}
