% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_centering_by_residuals.R
\name{getrsdls}
\alias{getrsdls}
\title{getrsdls}
\usage{
getrsdls(dat)
}
\arguments{
\item{dat}{}
}
\description{
This function used to actually use the residuals from a fit line around a particular person's data, but now we simply use the mean. Could go back to the previous version sometime.
}
