% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_actor_partner_cross.R
\name{makeEqual}
\alias{makeEqual}
\title{makeEqual}
\usage{
makeEqual(dat)
}
\arguments{
\item{dat}{}
}
\description{
this function takes a full dataframe, processes in groups by dyad, equalizes each of the partners data based on min and max times, and returns a new dataframe for the whole experiment population which is balanced.
}
\examples{
makeEqual(my.unbalanced.data.frame)
}
