% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_centering_by_residuals.R
\name{makeResidCol}
\alias{makeResidCol}
\title{makeResidCol}
\usage{
makeResidCol(dat)
}
\arguments{
\item{dat}{}
}
\description{
This function is used tot place teh residual column in the actual dataframe.
}
