% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDerivativesEstimate.R
\name{runDerivativesEstimate}
\alias{runDerivativesEstimate}
\title{runDerivativesEstimate}
\usage{
runDerivativesEstimate(deltaTime, theEmbed, theTau, dat_param)
}
\arguments{
\item{dat_param}{}
}
\description{
runDerivativesEstimate
}
