% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDerivativesEstimate.R
\name{runDerivativesEstimate}
\alias{runDerivativesEstimate}
\title{runDerivativesEstimate}
\usage{
runDerivativesEstimate(deltaTime, theTau, theEmbed, dat_param, est_info = F)
}
\arguments{
\item{dat_param}{}
}
\description{
Uses the Stephen Boker GLLA code to calculate 1st and 2nd derivatives for the person data that has been passed in at the moment.
}
